Moved patchDomainResourceWithNewRestartVersion to TestAction classs
